tune presentation of courses and videos
