+++ src/components/icons/rss.tsx
@@ -1,0 +1,24 @@
+import * as React from 'react'
+
+const RSSIcon: React.FunctionComponent<{className?: string}> = ({
+  className,
+}) => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      fill="none"
+      viewBox="0 0 24 24"
+      stroke="currentColor"
+      className={className}
+    >
+      <path
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        strokeWidth={2}
+        d="M6 5c7.18 0 13 5.82 13 13M6 11a7 7 0 017 7m-6 0a1 1 0 11-2 0 1 1 0 012 0z"
+      />
+    </svg>
+  )
+}
+
+export default RSSIcon
