+++ src/components/pages/courses/user-rating.tsx
@@ -16,10 +16,19 @@ const UserRating: React.FunctionComponent<{
         <StarIcon></StarIcon>
       </div> */}
       <div className="inline-flex flex-wrap items-center md:justify-start justify-center">
-        Rated <span className="ml-2 font-semibold">{rating.toFixed(1)}/5</span>
-        {' ・ '}
-        <span className="font-semibold mr-2">{count}</span> people completed
-        this {children} course
+        {rating > 0 && (
+          <>
+            Rated{' '}
+            <span className="ml-2 font-semibold">{rating.toFixed(1)}/5</span>
+          </>
+        )}
+        {rating > 0 && count > 0 && ' ・ '}
+        {count > 0 && (
+          <>
+            <span className="font-semibold mr-2">{count}</span> people completed
+            this {children} course
+          </>
+        )}
       </div>
     </div>
   )
